review: changed ksButton name to LinkButton
