+++ components/emergencyNumbers/EmergencyNumbers.tsx
@@ -9,7 +9,7 @@ import {
   Linking,
   ImageSourcePropType,
 } from 'react-native';
-import KsButton, { KsButtonProp } from '../shared/components/button/KsButton';
+import LinkButton, { KsButtonProp } from '../shared/components/button/LinkButton';
 import Card from '../shared/layout/Card';
 import { emergencies } from '../../assets/data/emergency';
 import { Emergency } from '../shared/model/emergency';
@@ -49,12 +49,12 @@ export default function EmergencyNumbers() {
               {/* Emergency service features */}
               {emergency.buttons.map((button: KsButtonProp, buttonIndex) => {
                 return (
-                  <KsButton backgroundColor="#FFFFFF" icon={button.icon} key={buttonIndex} link={button.link}>
+                  <LinkButton backgroundColor="#FFFFFF" icon={button.icon} key={buttonIndex} link={button.link}>
                     <View>
                       <Text style={styles.emergencyButtonLabel}>{button.label}</Text>
                       {!!button.description && <Text>{button.description}</Text>}
                     </View>
-                  </KsButton>
+                  </LinkButton>
                 );
               })}
             </View>
